Add gRPC stream to core/mutaitonserver
